requirements.txt
Adafruit-DHT==1.4.0
paho-mqtt==1.4.0